backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
pydantic-settings==2.4.0

openai>=1.42.0
numpy==1.26.4
pandas==2.2.2
duckdb==1.0.0
pdfplumber==0.11.0
tabulate==0.9.0
faiss-cpu==1.8.0.post1
python-multipart==0.0.9
boto3==1.34.131
sentence-transformers